split out `va_list` forwarding into its own test
